minor change to try to capture error msg when cleaning up node
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/DJob/trunk@9933 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ DistribJob/lib/perl/SgeNode.pm
@@ -107,9 +107,13 @@ sub cleanUp {
   if($self->{jobid} =~ /^(\d+)/){
     $jid = $1;
   }
-  my $errBase = basename($self->{script});
-  my $delCmd = "/bin/rm $errBase.?$jid > /dev/null 2>&1";
-  system($delCmd); 
+  if($self->{script}){
+    my $errBase = basename($self->{script});
+    my $delCmd = "/bin/rm $errBase.?$jid > /dev/null 2>&1";
+    system($delCmd); 
+  }else{
+    print "ERROR MSG for basename ... script name = '$self->{script}'\n";
+  }
 }
 
 1;
